requirements.txt
Flask==2.1.0
mysql-connector==2.2.9
Werkzeug==2.1.0
PyMySQL==1.1.0
Pillow==10.0.1





